rwsem_rt: fix compilation error on -Werror
Comparison between signed and unsigned variables causes compilation
error on some compilers when -Werror is enabled.

Bug 1905220

+++ include/linux/rwsem_rt.h
@@ -47,7 +47,7 @@ do {							\
 
 static inline int rwsem_is_locked(struct rw_semaphore *sem)
 {
-	return atomic_read(&sem->readers) != READER_BIAS;
+	return atomic_read(&sem->readers) != (int)READER_BIAS;
 }
 
 static inline int rwsem_is_contended(struct rw_semaphore *sem)
